# src/main/java/api/dto/FinishRequest.java
package api.dto;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.ToString;

@Getter
@NoArgsConstructor
@AllArgsConstructor
@ToString
@Deprecated
public class FinishRequest {

    @NotNull
    public Boolean status;

    public Object response;
}