# src/main/openapi/openapi.yaml
---
openapi: 3.0.3
info:
  title: quarkus-problem API
  version: 1.0-SNAPSHOT
tags:
- name: Callback endpoint
paths:
  /rest/callback/{taskName}/finish:
    post:
      tags:
      - Callback endpoint
      summary: "[ADMIN] Used by remote entity to report Task completion."
      parameters:
      - name: taskName
        in: path
        required: true
        schema:
          minLength: 1
          type: string
      - name: err
        in: query
        schema:
          allOf:
          - $ref: "#/components/schemas/ErrorOption"
          - default: PASS_ERROR
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinishRequest"
      responses:
        "200":
          description: Success with results
        "400":
          description: Invalid input parameters or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      deprecated: true
  /rest/callback/{taskName}/succeed:
    post:
      tags:
      - Callback endpoint
      summary: "[ADMIN] Used by remote entity to report successful Task completion."
      parameters:
      - name: taskName
        in: path
        required: true
        schema:
          minLength: 1
          type: string
      - name: err
        in: query
        schema:
          allOf:
          - $ref: "#/components/schemas/ErrorOption"
          - default: PASS_ERROR
      requestBody:
        content:
          application/json:
            schema: {}
      responses:
        "200":
          description: Success with results
        "400":
          description: Invalid input parameters or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorOption:
      enum:
      - GOOD
      - BAD
      type: string
    ErrorResponse:
      type: object
      properties:
        errorType:
          type: string
        errorMessage:
          type: string
        object: {}
        stackTrace:
          type: string
    FinishRequest:
      required:
      - status
      type: object
      properties:
        status:
          type: boolean
        response: {}
      deprecated: true